@@ -23,23 +23,23 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^0.19.1
+  flutter_svg: ^0.19.3
   smooth_page_indicator: ^0.2.0
-  flutter_riverpod: ^0.12.1
+  flutter_riverpod: ^0.12.4
 
-  pin_code_fields: ^6.0.2
-  flutter_screenutil: ^4.0.2+3
+  pin_code_fields: ^6.1.0
+  flutter_screenutil: ^4.0.4
   flutter_xlider: ^3.4.0
   flutter_typeahead: ^1.9.3
   autocomplete_textfield: ^1.7.3
-  carousel_slider: ^2.3.1
+  carousel_slider: ^2.3.4
   flutter_custom_switch: ^0.0.2
-  google_maps_flutter: ^1.1.1
-  geolocator: ^6.1.13
+  google_maps_flutter: ^1.2.0
+  geolocator: ^6.2.1
   geocoder: ^0.2.1
   mask_text_input_formatter: ^1.2.1
   firebase_core: ^0.7.0
-  firebase_auth: ^0.20.0+1
+  firebase_auth: ^0.20.1
   cloud_firestore: ^0.16.0
   flutter_otp: ^0.3.2
   shared_preferences: ^0.5.12+4
